metadata.jsonに書き出してからimportするようにしてみた
