+++ .gitignore
@@ -2,3 +2,4 @@
 node_modules
 yarn-error.log
 out
+lib/metadata.json
